converted some callback functions to async
