--- lib/helpers/createTransactionId.js
@@ -1,15 +1,0 @@
-'use strict';
-
-const Crypto = require('crypto');
-
-const createTransactionId = (callback) => {
-
-    setImmediate(() => {
-
-        const hash = Crypto.randomBytes(20).toString('hex');
-
-        callback(null, hash);
-    });
-};
-
-module.exports = createTransactionId;
