+++ lib/helpers/index.js
@@ -1,23 +1,23 @@
 'use strict';
 
 module.exports = {
-    defaultConfiguration   : require('./defaultConfiguration'),
-    createConnectionString : require('./createConnectionString'),
-    convertToBufferAsync   : require('./convertToBufferAsync'),
-    createTransactionId    : require('./createTransactionId'),
-    exponentialBackoff     : require('./exponentialBackoff'),
-    getPackageData         : require('./getPackageData'),
-    isMajorCompatible      : require('./isMajorCompatible'),
-    cleanObject            : require('./cleanObject'),
-    parsePayload           : require('./parsePayload'),
-    reduceRouteKey         : require('./reduceRouteKey'),
-    reduceCallback         : require('./reduceCallback'),
-    validateLoggerContract : require('./validateLoggerContract'),
-    validatePromiseContract: require('./validatePromiseContract'),
-    routeMatcher           : require('./routeMatcher'),
-    handlerMatcher         : require('./handlerMatcher'),
-    toPromise              : require('./toPromise'),
-    isString               : require('./isString'),
-    retryAsync             : require('./retryAsync'),
-    timeoutAsync           : require('./timeoutAsync')
+    defaultConfiguration        : require('./defaultConfiguration'),
+    createConnectionString      : require('./createConnectionString'),
+    convertToBufferAsync        : require('./convertToBufferAsync'),
+    createTransactionIdAsync    : require('./createTransactionIdAsync'),
+    exponentialBackoff          : require('./exponentialBackoff'),
+    getPackageData              : require('./getPackageData'),
+    isMajorCompatible           : require('./isMajorCompatible'),
+    cleanObject                 : require('./cleanObject'),
+    parsePayload                : require('./parsePayload'),
+    reduceRouteKey              : require('./reduceRouteKey'),
+    reduceCallback              : require('./reduceCallback'),
+    validateLoggerContract      : require('./validateLoggerContract'),
+    validatePromiseContract     : require('./validatePromiseContract'),
+    routeMatcher                : require('./routeMatcher'),
+    handlerMatcher              : require('./handlerMatcher'),
+    toPromise                   : require('./toPromise'),
+    isString                    : require('./isString'),
+    retryAsync                  : require('./retryAsync'),
+    timeoutAsync                : require('./timeoutAsync')
 };
