+++ lib/index.js
@@ -403,26 +403,27 @@ class BunnyBus extends EventEmitter{
         return result;
     }
 
-    // async send(message, queue, options) {
+    async send(message, queue, options) {
 
-    //     const routeKey = Helpers.reduceRouteKey(null, options, message);
-    //     const source = (options && options.source);
+        const routeKey = Helpers.reduceRouteKey(null, options, message);
+        const source = (options && options.source);
+        const convertedMessaged = await Helpers.convertToBufferAsync(message);
 
-    //     const channelContext = await this._autoBuildChannelContext(BunnyBus.MANAGEMENT_CHANNEL_NAME());
+        const channelContext = await this._autoBuildChannelContext(BunnyBus.MANAGEMENT_CHANNEL_NAME());
 
 
-    // }
-
-    send(message, queue, options, callback) {
-
-        callback = Helpers.reduceCallback(options, callback);
-
-        if (callback === undefined) {
-            return Helpers.toPromise($.promise, $._send, message, queue, options);
-        }
-
-        $._send(message, queue, options, callback);
     }
+
+    // send(message, queue, options, callback) {
+
+    //     callback = Helpers.reduceCallback(options, callback);
+
+    //     if (callback === undefined) {
+    //         return Helpers.toPromise($.promise, $._send, message, queue, options);
+    //     }
+
+    //     $._send(message, queue, options, callback);
+    // }
 
     _send(message, queue, options, callback) {
